auth middleware

+++ backend/src/controllers/notes.ts
@@ -2,14 +2,19 @@ import { NextFunction, RequestHandler, Response, Request } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { asserisDefined } from "../util/asserisDefined";
 
 export const getNotes: RequestHandler = async (
   req: Request,
   res: Response,
   next: NextFunction
 ) => {
+  const authenticatedUserId = req.session.userId;
+
   try {
-    const notes = await NoteModel.find().exec();
+    asserisDefined(authenticatedUserId);
+
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -18,14 +23,20 @@ export const getNotes: RequestHandler = async (
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    asserisDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId))
       throw createHttpError(400, "Invalid note id");
 
     const note = await NoteModel.findById(noteId).exec();
 
     if (!note) throw createHttpError(404, "Note not found");
+
+    if (!note.userId.equals(authenticatedUserId))
+      throw createHttpError(401, "You cannot access this note");
 
     res.status(200).json(note);
   } catch (error) {
@@ -45,13 +56,16 @@ export const createNote: RequestHandler<
   unknown
 > = async (req, res, next) => {
   const { title, text } = req.body;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    asserisDefined(authenticatedUserId);
     if (!title) {
       throw createHttpError(400, "Note must have a title");
     }
 
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -81,7 +95,11 @@ export const updateNote: RequestHandler<
   const newTitle = req.body.title;
   const newText = req.body.text;
 
+  const authenticatedUserId = req.session.userId;
+
   try {
+    asserisDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId))
       throw createHttpError(400, "Invalid note id");
 
@@ -93,13 +111,15 @@ export const updateNote: RequestHandler<
 
     if (!note) throw createHttpError(404, "Note not found");
 
+    if (!note.userId.equals(authenticatedUserId))
+      throw createHttpError(401, "You cannot access this note");
+
     note.title = newTitle;
     note.text = newText;
 
     const updatedNote = await note.save();
 
     res.status(200).json(updatedNote);
-
   } catch (error) {
     next(error);
   }
@@ -109,9 +129,17 @@ interface DeleteNoteParam {
   noteId: string;
 }
 
-export const deleteNote: RequestHandler<DeleteNoteParam, unknown, unknown, unknown> = async (req, res, next) => {
-  const {noteId} = req.params;
+export const deleteNote: RequestHandler<
+  DeleteNoteParam,
+  unknown,
+  unknown,
+  unknown
+> = async (req, res, next) => {
+  const { noteId } = req.params;
+  const authenticatedUserId = req.session.userId;
   try {
+    asserisDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId))
       throw createHttpError(400, "Invalid note id");
 
@@ -119,11 +147,13 @@ export const deleteNote: RequestHandler<DeleteNoteParam, unknown, unknown, unkno
 
     if (!note) throw createHttpError(404, "Note not found");
 
+    if (!note.userId.equals(authenticatedUserId))
+      throw createHttpError(401, "You cannot access this note");
+
     await note.deleteOne();
 
     res.sendStatus(204);
-    
   } catch (error) {
     next(error);
   }
-}
+};
